make tmp being a long it
To make it clean, we need to lookup s32int or whatever libc/posix defines it.

long is also architecture dependend, iirc.

and we need 32 bit.

+++ src/libcinit/cinit_get_svc_status.c
@@ -15,7 +15,7 @@
 
 int cinit_get_svc_status(char *name)
 {
-   int tmp
+   long int tmp
    int offset = 0;
    char *p, *answer;
